[improve] split layouts/GalleryHome
