+++ src/layouts/GalleryHome/components/LoadingLayout.tsx
@@ -1,0 +1,21 @@
+import React from 'react'
+import Loading from 'components/Loading'
+
+export default function LoadingLayout() {
+  return (
+    <div
+      style={{
+        position: 'fixed',
+        top: 0,
+        left: 0,
+        display: 'flex',
+        justifyContent: 'center',
+        alignItems: 'center',
+        width: '100vw',
+        height: '100vh'
+      }}
+    >
+      <Loading />
+    </div>
+  )
+}
